pyproject.toml
[project]
name = "ai-app-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "groq>=0.31.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.6",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
